Create a hook to determine the provider form schema
