+++ src/types/utils.ts
@@ -5,3 +5,7 @@ export type WithRequiredProperty<T, K extends keyof T> = T & {
 export type WithRequiredNonNullProperty<T, K extends keyof T> = T & {
     [P in K]-?: Exclude<T[P], null>;
 };
+
+export type EnumDictionary<E extends string | symbol | number, T> = {
+    [K in E]: T;
+};
